requirements.txt
asttokens==2.2.1
backcall==0.2.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.4
colorama==0.4.6
comm==0.1.2
cryptography==41.0.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock==3.9.0
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.26.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3
platformdirs==3.1.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==305
PyYAML==6.0.1
pyzmq==25.0.0
requests==2.31.0
setuptools==68.0.0
stack-data==0.6.2
sympy==1.11.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==1.26.18
wcwidth==0.2.6
wheel==0.41.2
win-inet-pton==1.1.0
